refactor: move endpoints to controller
